sharness: make `test_dir_is_empty` more robust
Summary:
The previous behavior could incorrectly pass given a directory with a
filename containing a newline. This patch changes the implementation to
use `ls -A`, which is specified by POSIX. The output should be empty
exactly if the directory is empty.

Test Plan:
The newly added unit test fails before this change and passes after it.

wchargin-branch: test-dir-is-empty-robust